refactor(net/rdr3): let each player patch file register itself
