--- code/components/gta-net-rdr3/src/players/PlayerPatches.cpp
@@ -1,50 +1,0 @@
-#include <StdInc.h>
-
-#include <Hooking.h>
-#include <MinHook.h>
-
-#include "PlayerPatches.h"
-
-static HookFunction hookFunction([]()
-{
-	ApplyPlayerDamageArrayPatches();
-	ApplyAnimSceneHandlerPatches();
-	ApplyAnimSceneArrayHandlerPatches();
-	ApplyPlayerCachePatches();
-	ApplyPlayerBandwidthPatches();
-	ApplyPlayerDamageTrackerPatches();
-	ApplyNetObjectPlayerIndexPatches();
-	ApplyNetPlayerMgrPatches();
-	ApplyPlayerIndexComparisons();
-	ApplyPlayerIterationPatches();
-	ApplyPedIntelligencePatches();
-	ApplyPlayerStackResizes();
-	ApplyBubbleJoinPatch();
-	ApplyPlayerBitsetPatches();
-	ApplyPlayerSyncDataPatch();
-	ApplyNodeDataSentPatch();
-	ApplySyncDataOwnerPatch();
-	ApplyScriptHandlerNodePatch();
-	ApplyGhostMaskPatch();
-	ApplyRemoteScriptInfoPatch();
-	ApplyScriptHandlerBitsetPatches();
-	ApplyCachedPlayerDataPatch();
-	ApplySyncMessageSequencePatch();
-	ApplyCachedPlayerLookupPatches();
-	ApplyPlayerFlagBuilderPatches();
-	ApplyRemainingBitsetPatches();
-	ApplyBitsetReadPatches();
-	ApplyActiveIndexPatches();
-	ApplyDisabledSubsystemPatches();
-	ApplyNetworkObjectMgrPatches();
-
-	MH_Initialize();
-
-	CreateDisabledSubsystemHooks();
-	CreatePlayerFocusHooks();
-	CreatePlayerBitsetHooks();
-	CreatePlayersNearPointHook();
-	CreateDisabledTelemetryHooks();
-
-	MH_EnableHook(MH_ALL_HOOKS);
-});
