--- code/components/gta-net-rdr3/src/players/PlayerPatches.h
@@ -1,38 +1,0 @@
-#pragma once
-
-void ApplyPlayerDamageArrayPatches();
-void ApplyAnimSceneHandlerPatches();
-void ApplyAnimSceneArrayHandlerPatches();
-void ApplyPlayerCachePatches();
-void ApplyPlayerBandwidthPatches();
-void ApplyPlayerDamageTrackerPatches();
-void ApplyNetObjectPlayerIndexPatches();
-void ApplyNetPlayerMgrPatches();
-void ApplyPlayerIndexComparisons();
-void ApplyPlayerIterationPatches();
-void ApplyPedIntelligencePatches();
-void ApplyPlayerStackResizes();
-void ApplyBubbleJoinPatch();
-void ApplyPlayerBitsetPatches();
-void ApplyPlayerSyncDataPatch();
-void ApplyNodeDataSentPatch();
-void ApplySyncDataOwnerPatch();
-void ApplyScriptHandlerNodePatch();
-void ApplyGhostMaskPatch();
-void ApplyRemoteScriptInfoPatch();
-void ApplyScriptHandlerBitsetPatches();
-void ApplyCachedPlayerDataPatch();
-void ApplySyncMessageSequencePatch();
-void ApplyCachedPlayerLookupPatches();
-void ApplyPlayerFlagBuilderPatches();
-void ApplyRemainingBitsetPatches();
-void ApplyBitsetReadPatches();
-void ApplyActiveIndexPatches();
-void ApplyDisabledSubsystemPatches();
-void ApplyNetworkObjectMgrPatches();
-
-void CreateDisabledSubsystemHooks();
-void CreatePlayerFocusHooks();
-void CreatePlayerBitsetHooks();
-void CreatePlayersNearPointHook();
-void CreateDisabledTelemetryHooks();
